+++ changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.1b1 (2021-Jan-14)
 -------------------
  * bugfix/coin not present in future not considered for buying
+ * Cutoff values can be set as mean-values
 
 1.1b0 (2020-Dec-xx)
 -------------------